docker: Reduce and reorder the steps to build container image.
The HTML and NGINX configuration are most likely to change compared to
the certificates and nginx apk update steps, so move those to the end.

We can also keep from having to have a dozen intermediate images by
copying all files that will go to /etc/nginx in one step.